Fix multiplayer scoreboard bugs
